Move aberration correction to the translate_from_to function
